+++ src/http/interceptors.ts
@@ -1,6 +1,34 @@
-import { AxiosResponse } from 'axios';
+import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
+import { delay } from '../utils';
 import { DATA_BASE } from '../constants';
+import { HttpResponseError } from './types';
 import { camelizeKeys, recursiveMap, recursiveRenameKeys } from './utils';
+
+export const retryInterceptor =
+  (axios: AxiosInstance) =>
+  async (error: HttpResponseError): Promise<unknown> => {
+    const { config } = error;
+    const { retry } = config;
+
+    // tries includes the initial try. So 5 tries equals 4 retries
+    if (
+      error?.isAxiosError &&
+      retry?.tries > retry?.current &&
+      retry?.retryCondition(error)
+    ) {
+      await delay(retry.retryTimeInMs);
+
+      return axios({
+        ...config,
+        retry: {
+          ...retry,
+          current: retry.current + 1,
+        },
+      } as AxiosRequestConfig);
+    }
+
+    return Promise.reject(error);
+  };
 
 export const camelizeResponseData = ({
   data,