+++ tests/unit/rql.test.ts
@@ -130,4 +130,18 @@ describe('rql string builder', () => {
       })`
     );
   });
+
+  it('should parse a string to rqlString', async () => {
+    const parsed = rqlBuilder().parse(
+      'or(like(name,bi,tfit),eq(),like(name,fitbit))'
+    );
+
+    expect(parsed).toBe('or(like(name,bi,tfit),eq(),like(name,fitbit))');
+  });
+
+  it('should throw an error on parsing invalid RQL to rqlString', async () => {
+    expect(() => {
+      rqlBuilder().parse('or((like(name,btfit))');
+    }).toThrowError();
+  });
 });